backend/requirements.txt
Flask
flask-cors
google-generativeai
pandas
openpyxl
werkzeug
python-docx
PyPDF2


